move outsider_events to own directory
