+++ scripts/events_module/outsider/outsider_events.py
